Renamed data-prepper-script to data-prepper-expression
