+++ config/checkstyle/checkstyle-suppressions.xml
@@ -11,5 +11,5 @@
 <suppressions>
     <!-- suppress all checks on files generated by ANTLR -->
     <suppress files="data-prepper-logstash-configuration[\\/]build[\\/]generated-src[\\/]antlr[\\/]main[\\/]*" checks="[a-zA-Z0-9]*"/>
-    <suppress files="data-prepper-script[\\/]build[\\/]generated-src[\\/]antlr[\\/]main[\\/]*" checks="[a-zA-Z0-9]*"/>
+    <suppress files="data-prepper-expression[\\/]build[\\/]generated-src[\\/]antlr[\\/]main[\\/]*" checks="[a-zA-Z0-9]*"/>
 </suppressions>
