+++ data-prepper-expression/src/main/antlr/DataPrepperScript.g4
@@ -6,7 +6,7 @@
 grammar DataPrepperScript;
 
 @header {
-    package org.opensearch.dataprepper.script.antlr;
+    package org.opensearch.dataprepper.expression.antlr;
 }
 
 fragment
