+++ data-prepper-expression/src/main/java/org/opensearch/dataprepper/expression/ScriptEngine.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.script;
+package org.opensearch.dataprepper.expression;
 
 import com.amazon.dataprepper.model.event.Event;
