+++ settings.gradle
@@ -49,5 +49,5 @@ include 'e2e-test:log'
 include 'data-prepper-test-common'
 include 'performance-test'
 include 'data-prepper-plugins:date-processor'
-include 'data-prepper-script'
+include 'data-prepper-expression'
